+++ openfisca_core/periods/instant_.py
@@ -2,13 +2,19 @@ from __future__ import annotations
 
 from typing import Tuple
 
-from dateutil import relativedelta
 import calendar
 import datetime
 import functools
 
+from dateutil import relativedelta
+
 from ._errors import DateUnitValueError, OffsetTypeError
 from ._units import DAY, MONTH, YEAR
+from .typing import Period
+
+from ._config import INSTANT_PATTERN
+
+from ._errors import InstantFormatError, InstantValueError
 
 
 class Instant(Tuple[int, int, int]):
@@ -198,3 +204,74 @@ class Instant(Tuple[int, int, int]):
             return Instant((date.year, date.month, date.day))
 
         return self
+
+    @staticmethod
+    def build(value: Any) -> Instant:
+        """Build a new instant, aka a triple of integers (year, month, day).
+
+        Args:
+            value: An ``instant-like`` object.
+
+        Returns:
+            None: When ``instant`` is None.
+            :obj:`.Instant`: Otherwise.
+
+        Raises:
+            InstantFormatError: When the arguments were invalid, like "2021-32-13".
+            InstantValueError: When the length is out of range.
+
+        Examples:
+            >>> build_instant(datetime.date(2021, 9, 16))
+            Instant((2021, 9, 16))
+
+            >>> build_instant(Instant((2021, 9, 16)))
+            Instant((2021, 9, 16))
+
+            >>> build_instant(Period(("year", Instant((2021, 9, 16)), 1)))
+            Instant((2021, 9, 16))
+
+            >>> build_instant("2021")
+            Instant((2021, 1, 1))
+
+            >>> build_instant(2021)
+            Instant((2021, 1, 1))
+
+            >>> build_instant((2021, 9))
+            Instant((2021, 9, 1))
+
+        """
+
+        if value is None:
+            return None
+
+        if isinstance(value, Instant):
+            return value
+
+        if isinstance(value, Period):
+            return value.start
+
+        if isinstance(value, str) and not INSTANT_PATTERN.match(value):
+            raise InstantFormatError(value)
+
+        if isinstance(value, str):
+            instant = tuple(int(fragment) for fragment in value.split('-', 2)[:3])
+
+        elif isinstance(value, datetime.date):
+            instant = value.year, value.month, value.day
+
+        elif isinstance(value, int):
+            instant = value,
+
+        elif isinstance(value, (tuple, list)) and not 1 <= len(value) <= 3:
+            raise InstantValueError(value)
+
+        else:
+            instant = tuple(value)
+
+        if len(instant) == 1:
+            return Instant((instant[0], 1, 1))
+
+        if len(instant) == 2:
+            return Instant((instant[0], instant[1], 1))
+
+        return Instant(instant)