+++ openfisca_core/taxbenefitsystems/tax_benefit_system.py
@@ -406,7 +406,7 @@ class TaxBenefitSystem:
             key = instant.start
 
         elif isinstance(instant, (str, int)):
-            key = periods.build_instant(instant)
+            key = periods.Instant.build(instant)
 
         else:
             msg = f"Expected an Instant (e.g. Instant((2017, 1, 1)) ). Got: {instant}."